Search view + pagination function - without checkbox value
