+++ index.html
@@ -8,11 +8,11 @@
 </head>
 <body>
     
-    <input type="text" id='value'>
-    <button id='button'>search</button>
-    <div id="result">
+    <input type="text" id='townSearch'>
+    <button id='townFind'>search</button>
+    <nav>
 
-    </div>
+    </nav>
     <div id="pages">
         
     </div>
